Wrote tests for all optimizers
Wrote a test that reads all optimizers and tries them out on a simple relaxation inversionrecovery problem.
I noticed that the initial scheme sometimes gets stuck if it has an ill conditioned jacobian? (which I dont really understand why it happens so for now I raise a value error in optimize).